Grant permissions to XDG directories
Grant permissions to XDG directories commonly used by Wine and Windows
programs. This should (hopefully) solve common user problems with
Protontricks lacking the permission to access some files and causing
"file not found" errors in the programs that are hard for the user to
pin down to missing Flatpak permissions.

Refs #25